@@ -22,15 +22,13 @@ class BarRailTabs extends StatefulWidget {
   final int initialTab;
   final List<BarRailTab> tabs;
   final List<Widget>? appBarActions;
-  final bool includeDrawer;
 
   const BarRailTabs(
       {Key? key,
       required this.title,
       required this.tabs,
       this.initialTab = 0,
-      this.appBarActions,
-      this.includeDrawer = true})
+      this.appBarActions})
       : super(key: key);
 
   @override
@@ -64,7 +62,6 @@ class _BarRailTabsState extends State<BarRailTabs> {
   Widget _buildPortrait(BuildContext context) {
     return Scaffold(
       appBar: _buildAppBar(),
-      drawer: _buildAppDrawer(),
       body: widget.tabs[_selectedTab].content,
       bottomNavigationBar: NavigationBar(
         selectedIndex: _selectedTab,
@@ -78,7 +75,6 @@ class _BarRailTabsState extends State<BarRailTabs> {
   Widget _buildLandscape(BuildContext context) {
     return Scaffold(
         appBar: _buildAppBar(),
-        drawer: _buildAppDrawer(),
         body: Row(
           children: [
             NavigationRail(
@@ -100,14 +96,6 @@ class _BarRailTabsState extends State<BarRailTabs> {
       title: Text(widget.title),
       actions: widget.appBarActions,
     );
-  }
-
-  Widget? _buildAppDrawer() {
-    if (widget.includeDrawer) {
-      return const DefaultDrawer();
-    }
-
-    return null;
   }
 
   List<T> _makeNavigationList<T>(
